+++ Tests/BlueTriangleTests/BlueTriangleTests.swift
@@ -7,11 +7,26 @@ final class BlueTriangleTests: XCTestCase {
     static var uploaderQueue: DispatchQueue = Mock.uploaderQueue
     static var onSendRequest: (Request) -> Void = { _ in }
 
+    static var timeIntervals: [TimeInterval] = []
+    static var timeIntervalProvider: () -> TimeInterval = {
+        timeIntervals.popLast() ?? 0
+    }
+
+    static var makePerformanceMonitor: () -> PerformanceMonitoring = {
+        PerformanceMonitorMock()
+    }
+
     override class func setUp() {
         super.setUp()
         BlueTriangle.configure { configuration in
+            configuration.timerConfiguration = .init(timeIntervalProvider: timeIntervalProvider)
             configuration.uploaderConfiguration = Mock.makeUploaderConfiguration(queue: uploaderQueue) { request in
                 onSendRequest(request)
+            }
+            configuration.performanceMonitorBuilder = PerformanceMonitorBuilder { sampleInterval in
+                {
+                    makePerformanceMonitor()
+                }
             }
         }
         BlueTriangle.prime()
@@ -20,6 +35,7 @@ final class BlueTriangleTests: XCTestCase {
     
     override func tearDown() {
         Self.onSendRequest = { _ in }
+        Self.makePerformanceMonitor = { PerformanceMonitorMock() }
         BlueTriangle.reset()
         super.tearDown()
     }
@@ -46,12 +62,11 @@ final class BlueTriangleTests: XCTestCase {
         let expectedInteractiveTime: TimeInterval = 1000
         let expectedEndTime: TimeInterval = 2000
 
-        var timeIntervals = [
+        Self.timeIntervals = [
             expectedEndTime,
             expectedInteractiveTime,
             expectedStartTime,
         ]
-        let timerConfiguration = Mock.makeTimerConfiguration { timeIntervals.popLast() ?? 0 }
 
         let requestExpectation = self.expectation(description: "Request sent")
         var request: Request!
@@ -79,7 +94,6 @@ final class BlueTriangleTests: XCTestCase {
             Mock.configureBlueTriangle(configuration: config)
             // Internal
             config.requestBuilder = requestBuilder
-            config.timerConfiguration = timerConfiguration
         }
 
         let timer = BlueTriangle.startTimer(page: Mock.page)
@@ -101,5 +115,59 @@ final class BlueTriangleTests: XCTestCase {
             osVersion: Device.osVersion)
 
         XCTAssertEqual(requestString, expectedString)
+    }
+}
+
+extension BlueTriangleTests {
+    @available(iOS 14.0, *)
+    func testDisplayLinkPerformanceMonitor() throws {
+        let requestExpectation = self.expectation(description: "Request sent")
+        var request: Request!
+        Self.onSendRequest = { req in
+            request = req
+            requestExpectation.fulfill()
+        }
+
+        var finishedTimer: BTTimer!
+        let requestBuilder = RequestBuilder { session, timer, purchaseConfirmation in
+            finishedTimer = timer
+            let model = TimerRequest(session: session,
+                                     page: timer.page,
+                                     timer: timer.pageTimeInterval,
+                                     purchaseConfirmation: purchaseConfirmation,
+                                     performanceReport: timer.performanceReport)
+
+            return try Request(method: .post,
+                               url: Constants.timerEndpoint,
+                               headers: nil,
+                               model: model)
+        }
+
+        var performanceMonitor: DisplayLinkPerformanceMonitor!
+        Self.makePerformanceMonitor = {
+            performanceMonitor = DisplayLinkPerformanceMonitor(minimumSampleInterval: 0.1,
+                                                               resourceUsage: ResourceUsage.self)
+            return performanceMonitor
+        }
+
+        BlueTriangle.configure { config in
+            Mock.configureBlueTriangle(configuration: config)
+            // Internal
+            config.requestBuilder = requestBuilder
+        }
+
+        // ViewController
+        let imageSize: CGSize = .init(width: 150, height: 150)
+        let delayStragegy: DelayGenerator.Strategy = .random((mean: 1, variation: 0.5))
+        let networkClient: NetworkClientMock = .makeClient(delayStrategy: delayStragegy,
+                                                           imageSize: imageSize)
+        let viewController = CollectionViewController(networkClient: networkClient)
+        viewController.viewDidLoad()
+
+        waitForExpectations(timeout: 10.0)
+
+        XCTAssertNotNil(finishedTimer)
+        print(performanceMonitor.measurements)
+        // ...
     }
 }
